+++ gpt2giga/routers/__init__.py
@@ -1,5 +1,5 @@
-from .anthropic_router import router as anthropic_router
-from .api_router import router as api_router
+from .anthropic import router as anthropic_router
+from .api import router as api_router
 from .logs_router import logs_api_router, logs_router
 from .system_router import system_router
