+++ tests/test_router/test_anthropic_router.py
@@ -9,7 +9,7 @@ from loguru import logger
 
 from gpt2giga.models.config import ProxyConfig
 from gpt2giga.protocol import ResponseProcessor
-from gpt2giga.routers.anthropic_router import (
+from gpt2giga.routers.anthropic import (
     _build_anthropic_response,
     _convert_anthropic_messages_to_openai,
     _convert_anthropic_tools_to_openai,
