+++ tests/test_router/test_router_endpoints.py
@@ -7,7 +7,7 @@ from loguru import logger
 
 from gpt2giga.models.config import ProxyConfig
 from gpt2giga.protocol import ResponseProcessor
-from gpt2giga.routers.api_router import router
+from gpt2giga.routers.api import router
 
 
 class MockResponse:
